Add "reset" to loader
Fixes #11 - still need to remove vegas, but this removed unusable parts and cleaned things up.

Because there are no longer IDs, it isn't possible know which items are being deleted. So we just delete them all. This update ensures caching still works across a reset because it doesn't remove the results data, so this is a leak if you're loading data that changes frequently.